Remove support left semi -> inner join
